+++ docs/Getting-Started/Debug.md
@@ -214,7 +214,7 @@ Set-PodeCurrentRunspaceName -Name 'Custom Runspace Name'
 
 This cmdlet sets a custom name for the runspace, making it easier to track during execution.
 
-#### Example
+#### Example: Setting a Custom Runspace Name in a Task
 
 Here’s an example that demonstrates how to set a custom runspace name in a Pode task:
 
@@ -240,7 +240,7 @@ Get-PodeCurrentRunspaceName
 
 This cmdlet returns the name of the current runspace, allowing for easier tracking and management in complex scenarios with multiple concurrent runspaces.
 
-#### Example
+#### Example: Outputting the Runspace Name in a Schedule
 
 Here’s an example that uses `Get-PodeCurrentRunspaceName` to output the runspace name during the execution of a schedule:
 
@@ -291,7 +291,7 @@ try {
 }
 catch {
     # Capture the dump with custom options
-    Invoke-PodeDump -ErrorRecord $_ -Format 'clixml' -Path 'C:\CustomDump' -Halt
+    Invoke-PodeDump -ErrorRecord $_ -Format 'clixml' -Path 'C:\CustomDump'
 }
 ```
 
@@ -299,7 +299,6 @@ In this example:
 
 - The memory dump is saved in CLIXML format instead of the default.
 - The dump file is saved in the specified directory (`C:\CustomDump`) instead of the default path.
-- The `-Halt` switch will terminate the application after the dump is saved.
 
 ### Using the Dump Feature
 
@@ -319,13 +318,72 @@ Start-PodeServer -EnableBreakpoints {
         }
         catch {
             # Invoke a memory dump when a critical error occurs
-            $_ | Invoke-PodeDump -Halt
+            $_ | Invoke-PodeDump
         }
     }
 }
 ```
 
-In this setup, if an error occurs in the route, `Invoke-PodeDump` is called, capturing the current state and halting the application if the `-Halt` switch is set.
+In this setup, if an error occurs in the route, `Invoke-PodeDump` is called, capturing the current state.
+
+### Dump File Content
+
+When a dump file is created using the `Invoke-PodeDump` function, it captures detailed information about the state of the Pode application. The captured information provides valuable insights into the internal state at the time of the error or diagnostic event. Below is a description of the content included in the dump file:
+
+#### **1. Timestamp**
+
+- The exact date and time when the dump was generated.
+- Format: ISO8601 (e.g., `2024-12-01T12:34:56`).
+
+#### **2. Memory Details**
+
+- Provides information about the memory usage of the Pode application at the time of the dump.
+- Captures metrics such as total memory allocated and other relevant details to analyze memory-related issues.
+
+#### **3. Script Context**
+
+- Contains details about the script execution context at the time of the dump.
+- Includes information such as the script path, parameters passed, and the execution environment.
+
+#### **4. Stack Trace**
+
+- A snapshot of the call stack, showing the sequence of function calls leading up to the dump.
+- Useful for identifying the source of an error or pinpointing the exact location in the code where the issue occurred.
+
+#### **5. Exception Details**
+
+- Contains information about any exceptions that triggered the dump.
+- Includes the exception type, message, and inner exception details (if available).
+
+#### **6. Scoped Variables**
+
+- Lists all variables currently in scope, including their names and values.
+- This information is useful for debugging issues related to variable states.
+
+#### **7. Runspace Pool Details**
+
+- Describes the state and configuration of each runspace pool within the Pode application.
+- Captures the following details for each runspace pool:
+  - **Pool Name**: The name assigned to the runspace pool.
+  - **State**: The current state of the pool (e.g., `Opened`, `Closing`).
+  - **Result**: The result of the last operation performed on the pool.
+  - **Instance ID**: A unique identifier for the runspace pool instance.
+  - **Is Disposed**: Indicates whether the pool has been disposed.
+  - **Runspace Pool State Info**: Detailed state information about the runspace pool.
+  - **Initial Session State**: The initial session state used to configure the pool.
+  - **Cleanup Interval**: The interval used to clean up the pool.
+  - **Availability**: The availability of the runspace pool for handling requests.
+  - **Thread Options**: The threading options configured for the pool.
+
+#### **8. Runspace Details**
+
+- Provides detailed information about each individual runspace within the application.
+- Captures the following information for each runspace:
+  - **Runspace ID**: A unique identifier for the runspace.
+  - **Name**: The name of the runspace, often reflecting its purpose (e.g., `Pode_Web_Listener_1`).
+  - **Scoped Variables**: A detailed list of variables scoped to this runspace, captured using the Pode debugger.
+  - **Initial Session State**: The initial session state of the runspace.
+  - **Runspace State Info**: Detailed state information about the runspace, such as `Opened`, `Busy`, or `Disconnected`.
 
 ### Dumping Options and Best Practices
 
@@ -337,4 +395,5 @@ In this setup, if an error occurs in the route, `Invoke-PodeDump` is called, cap
 - **Setting the Path**: Use a dedicated folder for dump files (e.g., `./Dump`) to keep diagnostic files organized. The default path in the configuration can be overridden at runtime if needed.
 
 With these configurations and usage practices, the memory dump feature in Pode can provide a powerful tool for diagnostics and debugging, capturing critical state information at the time of failure.
- 
+
+
