+++ src/Embedded/DebuggerHandler.cs
@@ -5,105 +5,93 @@ using System.Collections.ObjectModel;
 
 namespace Pode.Embedded
 {
-    public class DebuggerHandler
+    public class DebuggerHandler : IDisposable
     {
         // Collection to store variables collected during the debugging session
-        private static PSDataCollection<PSObject> variables = new PSDataCollection<PSObject>();
+        public PSDataCollection<PSObject> Variables { get; private set; } = new PSDataCollection<PSObject>();
 
         // Event handler for the DebuggerStop event
-        private static EventHandler<DebuggerStopEventArgs> debuggerStopHandler;
+        private EventHandler<DebuggerStopEventArgs> DebuggerStopHandler;
 
         // Flag to indicate whether the DebuggerStop event has been triggered
-        private static bool eventTriggered = false;
+        public bool IsEventTriggered { get; private set; } = false;
 
         // Flag to control whether variables should be collected during the DebuggerStop event
-        private static bool shouldCollectVariables = true;
+        private bool CollectVariables = true;
 
-        // Method to attach the DebuggerStop event handler to the runspace's debugger
-        public static void AttachDebugger(Runspace runspace, bool collectVariables = true)
+        // Runspace object to store the runspace that the debugger is attached to
+        private Runspace Runspace;
+
+        public DebuggerHandler(Runspace runspace, bool collectVariables = true)
         {
-            // Set the collection flag based on the parameter
-            shouldCollectVariables = collectVariables;
+            // Set the collection flag and Runspace object
+            CollectVariables = collectVariables;
+            Runspace = runspace;
 
             // Initialize the event handler with the OnDebuggerStop method
-            debuggerStopHandler = new EventHandler<DebuggerStopEventArgs>(OnDebuggerStop);
-
-            // Attach the event handler to the DebuggerStop event of the runspace's debugger
-            runspace.Debugger.DebuggerStop += debuggerStopHandler;
+            DebuggerStopHandler = new EventHandler<DebuggerStopEventArgs>(OnDebuggerStop);
+            Runspace.Debugger.DebuggerStop += DebuggerStopHandler;
         }
 
-        // Method to detach the DebuggerStop event handler from the runspace's debugger
-        public static void DetachDebugger(Runspace runspace)
+        // Method to detach the DebuggerStop event handler from the runspace's debugger, and general clean-up
+        public void Dispose()
         {
-            if (debuggerStopHandler != null)
-            {
-                // Remove the event handler to prevent further event handling
-                runspace.Debugger.DebuggerStop -= debuggerStopHandler;
+            IsEventTriggered = false;
 
-                // Set the handler to null to clean up
-                debuggerStopHandler = null;
+            // Remove the event handler to prevent further event handling
+            if (DebuggerStopHandler != default(EventHandler<DebuggerStopEventArgs>))
+            {
+                Runspace.Debugger.DebuggerStop -= DebuggerStopHandler;
+                DebuggerStopHandler = null;
             }
+
+            // Clean-up variables
+            Runspace = default(Runspace);
+            Variables.Clear();
+
+            // Garbage collection
+            GC.SuppressFinalize(this);
         }
 
         // Event handler method that gets called when the debugger stops
-        private static void OnDebuggerStop(object sender, DebuggerStopEventArgs args)
+        private void OnDebuggerStop(object sender, DebuggerStopEventArgs args)
         {
             // Set the eventTriggered flag to true
-            eventTriggered = true;
+            IsEventTriggered = true;
 
             // Cast the sender to a Debugger object
             var debugger = sender as Debugger;
-            if (debugger != null)
+            if (debugger == default(Debugger))
             {
-                // Enable step mode to allow for command execution during the debug stop
-                debugger.SetDebuggerStepMode(true);
+                return;
+            }
 
-                PSCommand command = new PSCommand();
+            // Enable step mode to allow for command execution during the debug stop
+            debugger.SetDebuggerStepMode(true);
 
-                if (shouldCollectVariables)
+            // Collect variables or hang the debugger
+            var command = new PSCommand();
+            command.AddCommand(CollectVariables
+                ? "Get-PodeDumpScopedVariable"
+                : "while($PodeContext.Server.Suspended) { Start-Sleep -Milliseconds 500 }");
+
+            // Execute the command within the debugger
+            var outputCollection = new PSDataCollection<PSObject>();
+            debugger.ProcessCommand(command, outputCollection);
+
+            // Add results to the variables collection if collecting variables
+            if (CollectVariables)
+            {
+                foreach (var output in outputCollection)
                 {
-                    // Collect variables
-                    command.AddCommand("Get-PodeDumpScopedVariable");
-                }
-                else
-                {
-                    // Execute a break
-                    command.AddCommand( "while( $PodeContext.Server.Suspended){ Start-sleep 1}");
-                }
-
-                // Create a collection to store the command output
-                PSDataCollection<PSObject> outputCollection = new PSDataCollection<PSObject>();
-
-                // Execute the command within the debugger
-                debugger.ProcessCommand(command, outputCollection);
-
-                // Add results to the variables collection if collecting variables
-                if (shouldCollectVariables)
-                {
-                    foreach (var output in outputCollection)
-                    {
-                        variables.Add(output);
-                    }
-                }
-                else
-                {
-                    // Ensure the debugger remains ready for further interaction
-                    debugger.SetDebuggerStepMode(true);
+                    Variables.Add(output);
                 }
             }
-        }
-
-
-        // Method to check if the DebuggerStop event has been triggered
-        public static bool IsEventTriggered()
-        {
-            return eventTriggered;
-        }
-
-        // Method to retrieve the collected variables
-        public static PSDataCollection<PSObject> GetVariables()
-        {
-            return variables;
+            else
+            {
+                // Ensure the debugger remains ready for further interaction
+                debugger.SetDebuggerStepMode(true);
+            }
         }
     }
 }
