+++ tests/unit/Helpers.Tests.ps1
@@ -1058,7 +1058,7 @@ Describe 'Get-PodeRelativePath' {
     }
 
     It 'Returns path for a relative path joined to default root' {
-        Get-PodeRelativePath -Path './path' -JoinRoot | Should -Be 'c:/./path'
+        Get-PodeRelativePath -Path './path' -JoinRoot | Should -Be 'c:/path'
     }
 
     It 'Returns resolved path for a relative path joined to default root when resolving' {
@@ -1072,7 +1072,7 @@ Describe 'Get-PodeRelativePath' {
     }
 
     It 'Returns path for a relative path joined to passed root' {
-        Get-PodeRelativePath -Path './path' -JoinRoot -RootPath 'e:/' | Should -Be 'e:/./path'
+        Get-PodeRelativePath -Path './path' -JoinRoot -RootPath 'e:/' | Should -Be 'e:/path'
     }
 
     It 'Throws error for path ot existing' {
